fix: add generateStaticParams and refactor MoodPage into client/server split for static export
